+++ CHANGELOG.md
@@ -1,14 +1,26 @@
-# Next: 1.2.0
+# Planned
 
 - Better "empty" states
 - Filtering issues table
 - Logging and a log viewer in settings
 
+# Next: 1.2.0
+
+### Added
+
+- Dark, Light, System theme preference persists in database and can be toggled from Settings or View menu ([#10](https://github.com/barney-campbell/jira-helper/pull/10))
+- Summary column in yesterday time tracking widget, widget now spans full dashboard width ([#12](https://github.com/barney-campbell/jira-helper/pull/12))
+- Display version in settings, check git tags for updates
+
 # 1.1.0
+
+### Added
 
 - Dashboard widget displaying time tracking from the previous working day
 - Kanban-style to do list with ability to link jira issues
 
 # 1.0.0
+
+### Added
 
 - Initial version ported from WPF app